Fixed docstring and behavior when arguments are lists instead of sets.
Before, the docstring for `simplify_los` said that `dc` is a set of
strings. This is in contrast to the default parameter value, which is
`[]`. I think the intended behavior is for both `ones` and `dc` to be
lists (as the "l" in the function name probably implies).

In `simplify`, `simplify_los` is called with sets as arguments. I
changed this so it's a list as intended.

Moreover, if we use `simplify_los` accordingly and pass lists as
arguments, we get an error because set union is used in `terms = ones | dc`.

I fixed this by creating sets from those lists there instead.